feat(node): Add `processSessionIntegration`
